Banking_new/requirements.txt
# ============================================
# 🏦 Banking Text-to-SQL Agent - Requirements
# ============================================

# Core Python packages
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0

# Database connectivity
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0

# AI/ML and LangChain ecosystem
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.20

# Vector database and embeddings
faiss-cpu>=1.7.4
openai>=1.3.0

# Web framework
streamlit>=1.28.0

# Data validation and serialization
pydantic>=2.5.0

# SQL parsing and validation
sqlparse>=0.4.4

# Logging and utilities
python-dateutil>=2.8.2

# Development and testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 